Support lerna-style repositories.
Previously, https urls to deep packages, such as those under
babel/packages/<package_name>, were being truncated to
babel/packages because this believed the branch name was actually
master/packages.

Unfortunately there is not really a way to know which component
of a github url is actually the branch, because branches can contain
slashes and they aren't escaped.

Despite this unfortunate reality, we can make a best-guess that
anything after master/ is actually a path, which fixes the vast
majority of lerna packages.

This should fix the sidebar on babel packages on npmjs.com.